Add DestinationFilter protocol to filter out unwanted messages
DestinationFilter sits right above the transport protocol to make sure
we only handle messages intended for us. We currently define "us" as a
message destined for our cluster and, if not multicast, for our
specific address.

It's a protocol so that it could be removed to emulate a promiscuous
mode, perhaps useful in a monitoring, metrics, or logging scenario.