Get Udp unicast and Channel memberAdded events working
We now maintain two Udp sockets - one for receiving unicast messages
and for sending all messages, and a separate one for receiving
multicast messages. This means we now have to try to bind the unicast
socket to a range of ports, since each Channel (and thus OS process)
needs its own port.

We start at MULTICAST_PORT and try up to MULTICAST_PORT + PORT_RANGE,
which means 7600 - 7700 given the defaults.

The demo was updated to greet each new member of the cluster as they
arrive using the new memberAdded event.