Add heartbeat failure detection and verification of suspect members
The Heartbeat protocol's failure detection is fairly basic at this
point. If we don't receive a heartbeat from a cluster member within a
certain interval, we mark it as suspect. That triggers the
VerifySuspect protocol to send a unicast message directly to that
member. If a reply isn't received within the timeout, the member is
marked as dead and removed from the cluster view.