Add Channel.stop, delegating down to each protocol in the stack.
This implementation is somewhat naive, since the stop function is
asynchronous all the way down, and it may take some time. During the
time it takes to stop everything, new messages may have been sent or
received. There are some simple checks in place at the transport layer
to ensure we don't try to send on a closed socket, but this is not
really enough to cover the range of possible scenarios.